Two postponed TODOs in recent PRs... (#1148)
* Move eval functions to mathics.eval.numbers.numbers
* Move "no_doc" after imports
